Add Track SID to track stats (#30)
* Add track SID to local and remote track stats

This commit adds track SID to base stats object and adds validation to
all existing stats integration tests.

* Prefix remote track stats fields and methods

* Remove reflection logic mapping track to participant in application

In order to build the stats UI in video app, track stats must be mapped
to participants. Previously, reflection was used to get the track id to
see if it matches a track id in a stats report. This method is no
longer necessary with track SIDs available in stats reports.

* Add changelog entry for track sid addition to BaseTrackStats