WIP - Read body
Try to come up with a better way to tear down the channel. The inifinite loop approach has made testing very complicated. Perhaps you could trigger the coroutine scope teardown after reading the last receieve value from the viewmodel. i.e. in the recieveOrNull block check if the value is empty, if it is then close the channel in the roommanager.