Replace the RooomManager RxJava PublishSubject with a Coroutine Channel (#153)
* Replace RxJava PublishSubject with Coroutine Channel

* WIP - Read body

Try to come up with a better way to tear down the channel. The inifinite loop approach has made testing very complicated. Perhaps you could trigger the coroutine scope teardown after reading the last receieve value from the viewmodel. i.e. in the recieveOrNull block check if the value is empty, if it is then close the channel in the roommanager.

* Immediately return Connecting RoomViewEvent

* Simplify channel cleanup logic

* Remove redundant ExperimentalCoroutinesApi annotation

* Remove redudant coroutineDispatcher property and add lint supress annotations

* Shutdown room manager in onCleared

* Put experimental annotation at top of view model

* Remove shutdown of channel since RoomManager is a singleton

* Remove experimental APIs